pyproject.toml
[tool.poetry]
name = "hamahiyogrambot"
version = "0.1.0"
description = ""
authors = ["root2116 <root6889@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
instascrape = "^0.1"
instaloader = "^4.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
